feat(platform): implement authentication check in layout and refactor to use client component
