+++ apps/platform/app/api/isAuthed/route.ts
@@ -1,0 +1,10 @@
+import { cookies } from 'next/headers'
+
+export async function GET() {
+  const cookieStore = await cookies()
+  const accessToken = cookieStore.get('access_token')
+
+  const hasAccessToken = !!accessToken?.value
+
+  return Response.json({ ok: hasAccessToken })
+}
